--- pixell/tests/__init__.py
@@ -1,3 +1,0 @@
-from .test_geom import *
-from .test_io import *
-from .test_pixell import *
