+++ tests/data/cutsky_geometry_scaled_20x_9.4_arcminute_pixel_shape_wcs_tuple.pkl
