+++ tests/test_geom.py
